Fixing Configuration section in README.
In multiple places the README references a Configuration section that
does not exist. The README should not reference a section that does
not exist. This section did exist but was removed in commit c1d49f6
"update readme and swagger".

The Configuration section has been added back with links to a full
configuration example and the implementation. This is easier than
trying to keep the README in sync with the implementation.

Fixes #680