Fix NoAssertion for crate component
For crate components, the license declared by clearlydefined summarizer
is the preferred choice.  See Commit "Fix handling of Rust crates with
license choices".

In the problematic case of
https://clearlydefined.io/definitions/crate/cratesio/-/mpmc/0.1.6,
scancode detected that the license is BSD-2-Clause-Views.  The
component's license retrieved from the pacakge repository is
"non-standard". This is convereted to NoAssertion in the clearlydefined
summary, and later set to be the package's license.

The solution is to move the logice of overriding crate license
information to aggregator and only override when the license declared
by clearlydefined summary is not "NOASSERTION".