Allow skip fetching license breakdown in get stat api
The call to get license breakdown is quite costly: to fetch the top 10 licenses for composer components in dev deployment, an aggregation query with sortByCount uses 2800 RU.  This license breakdown information retrieved is currently not used/displayed in UI.  Introduce a query parameter to skip fetching licence breakdown when this information is not needed.  The default value for the query parameter, withLicenses, is set to true, to be consistent with the current behaviour of the stat api.