Restrict PyPI coordinates name in PypiCoordinatesMapper
In most cases, the coordinates name is parsed from a path parameter or from a string representation of coordinates using '/' as a delimiter. In these cases, '/' will not be present in coordinates names. There is one use case where the name is passed as a property of coordinates in a JSON payload via PATCH /curations. This is the use case where the coordinates name can potentially contain '/'.

In PyPI, '/' is invalid in a package name (see https://peps.python.org/pep-0508/#names).  Add a check to restrict the coordinates name prior to querying the PyPI API.